Resolve ambiguity in git checkout (file vs branch)
When doing a `git checkout blah`, if branch `origin/blah`
exists we expect a new branch `blah` to be created, set to
track `origin/blah` and the working copy switched to it.

The problem comes where all this happens but a file `blah`
also exists. In that case, `git` will favor the interpretation
of reverting changes to `blah`.

So, to force `git` to understand `blah` as a branch, we add `--`
to each `git checkout` command, since the convention is that
all files come after the `--`, if given (so, if we give a `--`
and `blah` comes before, `blah` has to be a branch).

Fixes #5.