Refuse to merge work in progress branches
If someone calls a merge request "WIP - foo", then gitlab considers
it work in progress and will refuse to (accidentally) merge it.

Marge was not paying attention to this attribute, so she would try to
rebase and merge and, when merging fails, just panic and leave a message
saying "Something bad happened, check my logs".

Now it won't even try to rebase, and will leave a clear message.