Handle more gracefully a request becoming WIP
GitLab has a feature that will automatically mark a merge_request
as WIP on a push if the commit name starts with WIP, or is a fixup!
commit. So it would happen that if someone creates a merge request
with a commit starting with WIP, the request wouldn't be marked as
WIP, but the moment marge rebases, it would turn this on and then
the merge would fail with a 405.

In principle, this could also happen if after we rebased and while
we are waiting for CI, someone sets the branch to WIP. Now we will
handle this more gracefully by explaining what happened instead
of saying that something went wrong