Use fetchFromGitHub to pin nixpkgs
This means no network request will be needed after first download (the other way
to achieve that would be nix-1.12 and specify a sha256 in fetchTarball). The
fetchFromGitHub solution is a bit ugly in that it requires an existing version
of nixpkgs and not just `raw` nix, but that's probably more of theoretical
concern.