Fix trailfilter's error handling and its tests
We were testing for `pytest.raises(Exception)` and succeeding when
the test was actually failing for the wrong reasons (a `TypeError`).
So we should use the expected `suprocess.CalledProcessError` instead, which
signals the script ended with a non-zero return code.

Now, again, the script was crashing for the wrong reasons while die-ing:
`print` was trying trying to use a `str` when writing to a raw stderr buffer.

We fix this by using `stderr.write()` instead of print, and also extend the
tests so that they verify the error message is what is to be expected