Mark --use-merge-strategy as experimental
This feature can leave your branch with spurious merge commits
in case CI fails after processed by marge. To make it completely
useful, we'd likely need to identify and remove attempted merge-commits,
but we need actual experience in order to make a sensible implementation.

It is better to warn users about they experimental nature of this feature, then.