Fail to merge if changes already exist in target
It's possible that the changes in an MR already exist in the target
branch - whether this was merged in manually or someone else happened to
have a similar idea. If this happens, when we rebase the MR on the
target the commits will be dropped, essentially turning into a no-op.
This causes the filter-branch to fail as it expects some revisions to be
listed. Whilst this has the same end result, it's clearer if we explain
exactly what the issue is.