Try to continue batching as many MRs as possible
We'd previously fail the batch job and start a regular job if any merge
requests failed on rebasing. Instead, we can skip that one and continue
with the rest.  This way only the failing job is removed from the batch.

If there's not enough merge jobs for the batch then we'll continue with
a regular job as usual.