Add create-pipeline and job-regexp options
create-pipeline: If for some reason no valid pipeline is found for the
MR (none exists, outdated, etc.), Marge will create a new one using
the api. She will also leave a message, so she can know that she
already created a pipeline and will not do it again and again. Default
is false.

job-regexp: The CI may be configured to run different jobs depending
on the branch name, or pipeline source (push, trigger, api...). Marge
will check the jobs in the current pipeline against this regexp. If
any job matches the regexp, it's considered good and the status is
checked as usual. If none matches, it's not a valid pipeline and either
the merge will fail or a new pipeline will be created (see create-pipeline).

If using both create-pipeline and job-regexp, make sure the important
jobs are configured to run on "api" sources (see the "only" and "except"
options of GitLab CI), so that the pipeline created by Marge will
be valid.