Ignore GitLab server errors when looking for projects and MRs.
It is not uncommon for GitLab to have temporary errors (500 responses
or other). If this happens while processing a MR, the safe thing is
probably to abort. But if it happens while looking for projects or MRs,
it should be OK to just ignore the error and proceed with an empty
list (i.e., wait and rerun the loop).