Check MR is mergeable before accepting MR
It's useful to check that we are still able to merge the MR after it has
finished CI. Whilst we can check the error message afterwards, these are
not always clearer and we'd essentially be doing the same checks as
these anyway.

As an extra advantage, we also refuse to merge if someone has since
unassigned from Marge (something that people tend to do without
realising it has no effect after Marge has run it's preflight checks).