Add tests for forked merge requests
By default the tests run on a mock merge request with the same source
and target project. This leaves us without much coverage for cases where
we want to merge as a fork.

This change also restructures the tests to make better use of fixtures,
as this make's writing the code to test both types of merge request
easier.

Finally, we make sure that all mocks used have `autospec` set so that
we don't accidentally miss bugs where wrong methods are called, or wrong
arguments are passed.