Handle merging when source branch is master
Fixes an issue uncovered by #124 where Marge throws an assertion error
if the source/target branches are from the same project and the source
branch is master. We assume that merges will always be into master,
which is obviously not always the case; you may have a workflow
requiring you to merge from master into another branch (e.g. a
staging/production branch).