Add appropriate error if unresolved discussions on merge request
Note that this doesn't use the discussions API, as it's a rather
convoluted process that doesn't work entirely as expected (the resolved
field doesn't appear to exist, for instance). Instead, we just check if
this setting is enabled at the project level, and if we haven't assumed
it's a different error then this is the most likely cause.