Use a mock of git for testing jobs
Add the counterpart of Mocklab but for the
execution of git commands on the local repository.

This means that we no longer need to mock the
calls to `Job.update_from_target_branch_and_push`
and can actually cover all paths of execution.
In particular, we can now run all tests for
both `merge` and `rebase`.

The mock works by redefining the `Repo.git()` method,
so that it understands the set of git commands that
marge runs. It keeps an extremely simplified model
of a git repository (plus the remote repositories
to which it connects) and uses this to implement
each command

In addition, we use fixtures to try the cross product of
parameters (merge/rebase; add-tested yes/no;
add-part-of yes/no) which actually found minor errors
(fixed in next commit)