BUGFIX: Handling of protected branches was broken
In the first version of the code, `branch_rewritten` was meant
to be a flag that would indicate whether the code managed to
go past the `add_trailers()`, so that in the exception handlers
we could know in where the exception came from. But because the
name was ambiguous, in f956e701 it was apparently changed to
reflect whether the a rewrite was performed at all after the
rebase/merge. However, in some places of the error handling logic,
it was still being used as the former in some places.

So we rename the variables that are meant to be flagposts to
have a hopefully more clear meaning, and fix the condition when
we check if the branch is protected: we need to compare the final
commit to the initial commit, to find out if we were actually trying
to push new branch.

This bug was found by the refactoring of the single-job tests.