Only fetch projects with min access level
If we're on an appropriate GitLab version, that is. GitLab doesn't
return the right permissions for projects belonging to the user if the
project is in a subgroup, and Marge is set as a member of the top-level
group[1]. This is rather annoying as Marge will then complain that she
doesn't have permissions for a project when she does.

A new field available in GitLab version >= 11.2 allows us to only fetch
projects for which we have a certain access level. This returns the
right projects even with nested groups, so we can prefer this method if
it is available.