Fix accessing permissions when using min_access_level
We fetch projects using the `min_access_level` field to bypass a bug in
GitLab causing no permissions to be provided in some cases. When this
happens, retrieving the access level is not possible because of the same
bug meaning no permissions are available.

To workaround this, we can add a fallback permission with developer
access, since at this point we know the projects we have fetched have at
least this permission.