By default, ignore archived projects.
This commit adds a new CLI argument (--include-archived) and ENV
VAR (MARGE_INCLUDE_ARCHIVED) which defaults to False. If True,
Marge will process archived projects (which cannot have their MRs
merged) as well as live projects (which can).

The API for Project.fetch_all_mine() has changed to include an
optional argument include_archived which matches the CLI arg.

The current default behaviour is for Marge to process all projects,
which wastes time and resources, linearly with the number of
archived projects.