Fetch pipelines by merge request instead of branch
... if possible.

Since we operate on merge requests this makes more sense, but the
feature was only available from version 10.5.

An annoyance of this route is that it doesn't allow ordering or sorting,
which we use to determine the latest pipeline by. However it's very
simple to just do this ourselves; the only downside is this is no longer
customizable, but nothing in the codebase appears to use this feature
anyway.

The batch job tries cancel pipelines by their status, which we can't
filter by for this route (unless we do it manually as well) so I've left
that using the old method. This can mean some detached pipelines may
continue running after a batch job has completed, but this seems like a
very minor issue.