* Add new choice assigned_at to option merge_order
Previously the option merge_order had two choices created_at and updated_at.
This is problematic in large projects as both options does not allow
the maintainer to control the order of merge requests.

This commit adds a new choice assigned_at to option merge_order as
maintainers in large project often expect the order of the merge requests
related to order in which the merge requests did got assigned to the
marge-bot user.

Signed-off-by: Konrad Weihmann <konrad.weihmann@mbition.io>