Use target branch name in batch merge to allow multiple instances
A project can use multiple instances of marge bots assigned to
different target branches. This allows dealing with different
release branches in parallel.

However batches could not be used on multiple instances because
they conflicted on the name of the branch created for the batch.
By naming this branch after the target branch, then multiple
batch branches can exist.