marge.app: print stdout/stderr for Exceptions
This fixes the following problem: bot executes various commands with the
`timeout` arg, and may throw a TimeoutExpired exception.  However,
printing TimeoutExpired exception won't print stdout and stderr.

Due to absence of stdout/stderr anywhere in the output this confuses and
looks like the command didn't manage to get too far in execution. Like
for example, as if the remote server didn't respond, making a `git
clone` to time out.

Signed-off-by: Konstantin Kharlamov <Hi-Angel@yandex.ru>